refactor: move model classes into model package
