+++ src/main/java/systems/boos/traindelays/CommandLineInterface.java
@@ -4,6 +4,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
+import systems.boos.traindelays.model.Timetable;
 
 import java.time.ZoneId;
 import java.time.format.DateTimeFormatter;
