+++ src/main/java/systems/boos/traindelays/model/Event.java
@@ -1,4 +1,4 @@
-package systems.boos.traindelays;
+package systems.boos.traindelays.model;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
